[Rewards] Add support for "section" and "order" on UI cards
