+++ components/brave_rewards/core/engine/endpoints/brave/get_ui_cards_unittest.cc
@@ -41,6 +41,8 @@ TEST_F(RewardsGetUICardsTest, ExpectedResponse) {
     }],
     "partner-promo-card": {
       "title": "$card-title",
+      "section": "explore",
+      "order": 1,
       "items": [{
         "title": "$title",
         "description": "$description",
@@ -56,6 +58,8 @@ TEST_F(RewardsGetUICardsTest, ExpectedResponse) {
   auto& community_card = result->at(0);
   EXPECT_EQ(community_card->name, "community-card");
   EXPECT_EQ(community_card->title, "");
+  EXPECT_EQ(community_card->section, "");
+  EXPECT_EQ(community_card->order, 0);
   ASSERT_EQ(community_card->items.size(), static_cast<size_t>(1));
   EXPECT_EQ(community_card->items[0]->title, "$title");
   EXPECT_EQ(community_card->items[0]->description, "$description");
@@ -65,6 +69,8 @@ TEST_F(RewardsGetUICardsTest, ExpectedResponse) {
   auto& promo_card = result->at(1);
   EXPECT_EQ(promo_card->name, "partner-promo-card");
   EXPECT_EQ(promo_card->title, "$card-title");
+  EXPECT_EQ(promo_card->section, "explore");
+  EXPECT_EQ(promo_card->order, 1);
   ASSERT_EQ(promo_card->items.size(), static_cast<size_t>(1));
   EXPECT_EQ(promo_card->items[0]->title, "$title");
   EXPECT_EQ(promo_card->items[0]->description, "$description");
