+++ components/brave_rewards/core/mojom/rewards.mojom
@@ -365,5 +365,7 @@ struct UICardItem {
 struct UICard {
   string name;
   string title;
+  string section;
+  int32 order;
   array<UICardItem> items;
 };
