+++ components/brave_rewards/resources/rewards_page/lib/app_state.ts
@@ -137,6 +137,8 @@ export interface UICardItem {
 export interface UICard {
   name: string
   title: string
+  section: string
+  order: number
   items: UICardItem[]
 }
